fixed issue with Big*Int operator precedence and implicit conversion

+++ test/unit/lexer_number.cc
@@ -23,7 +23,7 @@ TEST(LexerNumberTest, Int0)
   token_t* token = lexer_next(lexer);
   ASSERT_NE((void*)NULL, token);
   ASSERT_EQ(TK_INT, token_get_id(token));
-  ASSERT_EQ((__uint128_t)0, token_int(token));
+  ASSERT_EQ(0, token_int(token));
   token_free(token);
 
   lexer_close(lexer);
@@ -41,7 +41,7 @@ TEST(LexerNumberTest, Int0Finishes)
   token_t* token = lexer_next(lexer);
   ASSERT_NE((void*)NULL, token);
   ASSERT_EQ(TK_INT, token_get_id(token));
-  ASSERT_EQ((__uint128_t)0, token_int(token));
+  ASSERT_EQ(0, token_int(token));
   token_free(token);
 
   token = lexer_next(lexer);
@@ -64,7 +64,7 @@ TEST(LexerNumberTest, IntDecimal)
   token_t* token = lexer_next(lexer);
   ASSERT_NE((void*)NULL, token);
   ASSERT_EQ(TK_INT, token_get_id(token));
-  ASSERT_EQ((__uint128_t)12345, token_int(token));
+  ASSERT_EQ(12345, token_int(token));
   token_free(token);
 
   lexer_close(lexer);
@@ -82,7 +82,7 @@ TEST(LexerNumberTest, IntDecimalDot)
   token_t* token = lexer_next(lexer);
   ASSERT_NE((void*)NULL, token);
   ASSERT_EQ(TK_INT, token_get_id(token));
-  ASSERT_EQ((__uint128_t)12345, token_int(token));
+  ASSERT_EQ(12345, token_int(token));
   token_free(token);
 
   token = lexer_next(lexer);
@@ -122,7 +122,7 @@ TEST(LexerNumberTest, IntBinary)
   token_t* token = lexer_next(lexer);
   ASSERT_NE((void*)NULL, token);
   ASSERT_EQ(TK_INT, token_get_id(token));
-  ASSERT_EQ((__uint128_t)20, token_int(token));
+  ASSERT_EQ(20, token_int(token));
   token_free(token);
 
   lexer_close(lexer);
@@ -226,7 +226,7 @@ TEST(LexerNumberTest, IntHexNoOverflow)
   token_t* token = lexer_next(lexer);
   ASSERT_NE((void*)NULL, token);
   ASSERT_EQ(TK_INT, token_get_id(token));
-  ASSERT_EQ((__uint128_t)-1, token_int(token));
+  ASSERT_EQ(-1, token_int(token));
   token_free(token);
 
   lexer_close(lexer);
@@ -278,7 +278,7 @@ TEST(LexerNumberTest, IntNoOctal)
   token_t* token = lexer_next(lexer);
   ASSERT_NE((void*)NULL, token);
   ASSERT_EQ(TK_INT, token_get_id(token));
-  ASSERT_EQ((__uint128_t)100, token_int(token));
+  ASSERT_EQ(100, token_int(token));
   token_free(token);
 
   lexer_close(lexer);
